+++ src/CS146/Graph.java
@@ -3,7 +3,8 @@ package CS146;
 import java.util.*;
 
 /**
- * Represents a graph data structure.
+ * Represents a graph of locations
+ * ArrayList of Nodes
  */
 public class Graph {
 
@@ -11,10 +12,20 @@ public class Graph {
     private Node sourceNode;
     private Node destinationNode;
 
+    /**
+     * Constructor
+     */
     public Graph() {
         vertices = new ArrayList<>();
     }
 
+    /**
+     * Dijkstra's algorithm.
+     * Worse Case: O(E + VlogV)
+     * @param source the starting point
+     * @param destination the ending point
+     * @return a list of strings that represent the path
+     */
     public List<String> shortestpath(Node source, Node destination) {
     	sourceNode = source;
     	destinationNode = destination;
@@ -67,10 +78,18 @@ public class Graph {
         return result;
     }
 
-	public ArrayList<Node> getVertice() {
+    /**
+     * Return the array list of vertices
+     * @return
+     */
+	public ArrayList<Node> getVertices() {
 		return vertices;
 	}
 
+    /**
+     * Adds a vertice to the graph.
+     * @param node
+     */
     public void addVertice(Node node) {
         vertices.add(node);
     }
